+++ R/check_filename.R
@@ -2,6 +2,7 @@
 #' @inheritParams read_checklist
 #' @export
 #' @importFrom git2r in_repository ls_tree
+#' @family both
 check_filename <- function(x = ".") {
   if (!inherits(x, "Checklist") || !"checklist" %in% x$get_checked) {
     x <- read_checklist(x = x)
@@ -48,6 +49,7 @@ Failing folder: `%s`",
         "\\.[a-zA-Z]+ignore", "\\.Rprofile",
         "DESCRIPTION", "NAMESPACE",
         "README\\.R?md", "NEWS\\.md",
+        "CODE_OF_CONDUCT.md", "CONTRIBUTING.md", "LICENSE.md",
         "Dockerfile"
       ),
       collapse = "|"